Fix a race condition while server shutdown
What happens is, if we use the same context to
cancel the redis poll and the workers, the order
is not granteed, hence a job can be polled while all workers
are already down. which means a job will be dequeued and never
processes.

In production this can cause clients to block forever waiting for
a job that will never get processes.

The fix is to make sure the redis poll is getting canceled first
and that there are no more job runnings, before stopping the workers
and finally the server loop.